refactor(sandbox): extract tls_intercept's inline test module to tests.rs
Matches the merged sibling convention (ca/tests.rs,
credential_firewall/tests.rs) now that tls_intercept.rs itself has grown
past the point of comfortably carrying its tests inline. Done last, after
the new malformed-material/invalid-host/allowlist tests landed, so they
move with the extraction instead of racing it.

The escape-hatch ratchet (reborn_tls_verification_escape_hatches.rs) had
one test, the_scan_exempts_tls_intercepts_own_test_module, that pinned
truncate_at_inline_test_module's behavior specifically against
tls_intercept.rs's (now-removed) inline mod. Retargeted that pin to
scope_key.rs, which still keeps its tests inline, and added a companion
test asserting tls_intercept.rs's extraction is picked up by the
is_standalone_test_file(*/tests.rs) exclusion instead — proving both
halves of the mechanism this ratchet depends on stay covered.
